--- src/Containers/GearListBoard/GearListBoardCard/index.js
@@ -1,3 +1,0 @@
-import GearListBoardCard from './GearListBoardCard';
-
-export default GearListBoardCard;
