--- src/Containers/GearListBoard/GearListBoardHeader/index.js
@@ -1,3 +1,0 @@
-import GearListBoardHeader from './GearListBoardHeader';
-
-export default GearListBoardHeader;
