--- src/Containers/Navigation/LoggedInMenu/index.js
@@ -1,3 +1,0 @@
-import LoggedInMenu from './LoggedInMenu';
-
-export default LoggedInMenu;
