--- src/Containers/Navigation/LoggedInSideMenu/index.js
@@ -1,3 +1,0 @@
-import LoggedInSideMenu from './LoggedInSideMenu';
-
-export default LoggedInSideMenu;
