--- src/Containers/Navigation/LoggedOutMenu/index.js
@@ -1,3 +1,0 @@
-import LoggedOutMenu from './LoggedOutMenu';
-
-export default LoggedOutMenu;
