--- src/Containers/Navigation/LoggedOutSideMenu/index.js
@@ -1,3 +1,0 @@
-import LoggedOutSideMenu from './LoggedOutSideMenu';
-
-export default LoggedOutSideMenu;
