+++ src/Containers/Navigation/components/LoggedOutSideMenu/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import avatar from '../../../data/images/avatar.png';
+import avatar from '../../../../data/images/avatar.png';
 
 const LoggedOutSideMenu = () => {
   return (
