+++ src/components/FormFields/RenderInput/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const RenderInput = React.forwardRef(({ type = 'text', ...rest }, ref) => {
+  return <input type={type} {...rest} ref={ref} />;
+});
+
+export default RenderInput;
